requirements.txt
# ============================================
# CMS Backend - Heroku Optimized (CPU-Only)
# ============================================
# Last updated: 2026-02-27
# Using: ONNX Runtime + MobileFaceNet for lightweight face recognition
# Total ML footprint: ~20MB (vs ~400MB with PyTorch)

# --- Core Framework ---
pydantic==1.10.12
fastapi==0.103.2
uvicorn==0.23.2
starlette==0.27.0

# --- Production Server ---
gunicorn==21.2.0

# --- Database ---
pymongo==4.15.4

# --- Authentication ---
python-jose==3.3.0
passlib==1.7.4

# --- File Handling ---
python-multipart==0.0.6
python-dotenv==1.0.1
openpyxl==3.1.2

# --- Core Utilities ---
typing-extensions==4.12.0
Pillow==10.0.1
numpy==1.26.4  # Fixed to 1.x for OpenCV compatibility
requests==2.31.0
httpx>=0.24.0

# --- PDF/Reports ---
reportlab>=4.0.0
matplotlib>=3.8.0

# --- System Monitoring ---
psutil>=5.9.0

# --- Image Processing (headless = smaller, no GUI deps) ---
opencv-python-headless==4.8.0.76

# ============================================
# FACE RECOGNITION - MOBILEFACENET (LIGHTWEIGHT)
# ============================================
# ONNX Runtime CPU (~15MB) - Much lighter than PyTorch (~280MB)
onnxruntime==1.17.0

# --- Cryptography / password hashing ---
# Required for bcrypt password hashing used in auth
bcrypt==4.0.1

# --- DNS for MongoDB+SRV ---
# Required when using mongodb+srv URIs
dnspython==2.4.2

# ============================================
# MODEL: MobileFaceNet (downloaded at runtime)
# - Size: ~4MB (vs FaceNet's 107MB)
# - Accuracy: 99.28% on LFW (vs 99.63% for FaceNet)
# - Speed: 4x faster inference
# ============================================

# ============================================
# REMOVED (to reduce slug size):
# - torch (~280MB) - replaced with onnxruntime (~15MB)
# - torchvision (~50MB) - not needed
# - facenet-pytorch (~107MB model) - replaced with MobileFaceNet (~4MB)
# - tensorflow, deepface, tf-keras - not needed
# ============================================
